model/requirements.txt
torch
torchvision
numpy
pillow
requests
duckduckgo-search
ddgs
tqdm